migrated to sequelize

+++ src/utils.ts
@@ -3,10 +3,12 @@ import { Application } from './declarations';
 import { buildFakeContractDefinitions } from './services/contract-definitions/faker'
 import { ContractDefinition } from './services/contract-definitions/contract-definitions.class';
 
-
-export const removeAllContractDefinitions = async (app: Application) => {
+export const removeAllContractDefinitionsIfExists = async (app: Application) => {
     const contractDefinitionsService = app.service('contract-definitions');
-    await contractDefinitionsService.remove(null)
+    const existingContractDefinitions = await contractDefinitionsService.find()
+    if (existingContractDefinitions.total > 0) {
+        await contractDefinitionsService.remove(null)
+    }
 }
 
 export const addContractDefinitionsFromFaker = async (app: Application) => {
@@ -16,4 +18,5 @@ export const addContractDefinitionsFromFaker = async (app: Application) => {
     const contractDefinitions2: ContractDefinition = contractDefinitions[1]
     await contractDefinitionsService.create(contractDefinitions1)
     await contractDefinitionsService.create(contractDefinitions2)
+
 }
